refactor: change slider logic
