+++ src/pages/detailedProductPage/DetailedProductPage.tsx
@@ -4,8 +4,8 @@ import { useLocation } from 'react-router-dom';
 
 import './DetailedProductPage.css';
 
+import { ProductSlider } from '../../components/DetailedProduct/ProductSlider';
 import { SingleImage } from '../../components/DetailedProduct/SingleImage';
-import { MySlider } from '../../components/DetailedProduct/Slider';
 import Header from '../../components/Header/Header';
 import ProductRepository from '../../services/ProductRepository';
 
@@ -44,7 +44,7 @@ const DetailedProductPage = () => {
 
   const ShowImage = () => {
     if (images.length > 1) {
-      return <MySlider slides={images}></MySlider>;
+      return <ProductSlider slides={images}></ProductSlider>;
     } else if (images.length === 1) {
       return <SingleImage imageData={images[0]} />;
     } else {
